refactor: run prettier

+++ contracts/operators/Synthetix/SynthetixOperator.sol
@@ -12,10 +12,13 @@ import "../../interfaces/IOperatorSelector.sol";
 
 /// @title The synthetix trading operator to execute a swap
 contract SynthetixOperator is ISynthetixOperator, IOperatorSelector {
-
     /// @dev Deploy with the storage contract
     constructor(IAddressResolver synthetixResolver) {
-        address synthetixStorage = Create2.deploy(0, bytes32("nested.synthetix.operator"), type(SynthetixStorage).creationCode);
+        address synthetixStorage = Create2.deploy(
+            0,
+            bytes32("nested.synthetix.operator"),
+            type(SynthetixStorage).creationCode
+        );
         SynthetixStorage(synthetixStorage).updateSynthetixResolver(synthetixResolver);
         SynthetixStorage(synthetixStorage).transferOwnership(msg.sender);
     }
@@ -24,19 +27,25 @@ contract SynthetixOperator is ISynthetixOperator, IOperatorSelector {
     function commitAndRevert(
         address own,
         bytes32 sourceCurrencyKey,
-        uint sourceAmount,
+        uint256 sourceAmount,
         bytes32 destinationCurrencyKey
     ) external override returns (uint256[] memory amounts, address[] memory tokens) {
-        amounts = new uint[](1);
+        amounts = new uint256[](1);
         tokens = new address[](1);
 
         IAddressResolver addressResolver = IAddressResolver(SynthetixStorage(storageAddress(own)).synthetixResolver());
         ISynthetix synthetix = ISynthetix(addressResolver.getAddress("Synthetix"));
 
-        require(address(synthetix) != address(0), "SynthetixOperator::commitAndRevert: Synthetix is missing from Synthetix resolver");
+        require(
+            address(synthetix) != address(0),
+            "SynthetixOperator::commitAndRevert: Synthetix is missing from Synthetix resolver"
+        );
 
         // This check is what synthetix.exchange() will perform, added here for explicitness
-        require(!synthetix.isWaitingPeriod(sourceCurrencyKey), "SynthetixOperator::commitAndRevert: Cannot exchange during the waiting period");
+        require(
+            !synthetix.isWaitingPeriod(sourceCurrencyKey),
+            "SynthetixOperator::commitAndRevert: Cannot exchange during the waiting period"
+        );
 
         // TODO approval ?
 
@@ -50,9 +59,13 @@ contract SynthetixOperator is ISynthetixOperator, IOperatorSelector {
     /// @notice Return the operator storage address
     /// @param own the operator address to build the storage address in delegatecall
     function storageAddress(address own) public pure returns (address) {
-        bytes32 _data =
-        keccak256(
-            abi.encodePacked(bytes1(0xff), own, bytes32("nested.synthetix.operator"), keccak256(type(SynthetixStorage).creationCode))
+        bytes32 _data = keccak256(
+            abi.encodePacked(
+                bytes1(0xff),
+                own,
+                bytes32("nested.synthetix.operator"),
+                keccak256(type(SynthetixStorage).creationCode)
+            )
         );
         return address(uint160(uint256(_data)));
     }
